requirements.txt
streamlit
pandas
matplotlib
bcrypt
python-dateutil
